Move Docker REST API to its own file
Paving the way for what's coming next (Team using the same code)
