+++ src/RobotManager.cpp
@@ -86,7 +86,7 @@ void RobotManager::startContainers() {
 
     for (std::shared_ptr<Robot> r : robots_) {
         r->container = std::make_unique<Container>("CIRCUS_" + r->name + "_container");
-        r->container->create(r, image, binds);  // this is for the next commit, don't mind it for now
+        r->container->create(r, image, binds);
         r->container->start();
     }
 }
